Restructure package into tidier folders
