+++ packages/react-dialog-async/src/DialogOutlet/index.ts
@@ -1,0 +1,1 @@
+export { default as DialogOutlet } from './DialogOutlet';
