+++ packages/react-dialog-async/src/DialogProvider/index.ts
@@ -1,0 +1,2 @@
+export { default as DialogProvider } from './DialogProvider';
+export { type DialogProviderProps } from './types';
