+++ packages/react-dialog-async/src/DialogProvider/types.ts
@@ -1,0 +1,8 @@
+import { PropsWithChildren } from 'react';
+
+export interface DialogProviderProps extends PropsWithChildren {
+  /**
+   * The default delay in milliseconds to wait before unmounting a dialog after it's closed.
+   */
+  defaultUnmountDelayInMs?: number;
+}
