+++ packages/react-dialog-async/src/context/DialogContext.tsx
@@ -1,11 +1,11 @@
 import { createContext } from 'react';
-import { DialogComponent } from './types';
+import { AsyncDialogComponent } from '../types';
 
 export interface dialogContextState {
   show: (
     dialogId: string,
     hash: number,
-    dialog: DialogComponent<any, any>,
+    dialog: AsyncDialogComponent<any, any>,
     data: unknown,
     unmountDelay?: number,
   ) => Promise<any>;
