+++ packages/react-dialog-async/src/context/IndividualDialogContext.tsx
@@ -1,5 +1,5 @@
 import { createContext, useContext } from 'react';
-import { AsyncDialogProps } from './types';
+import { AsyncDialogProps } from '../types';
 
 interface DialogContextValue<D = any, R = any> extends AsyncDialogProps<D, R> {
   isInsideDialogContext: boolean;
@@ -18,7 +18,10 @@ const IndividualDialogContext = createContext<DialogContextValue>(
   defaultDialogContextValue,
 );
 
-export const useDialogContext = <D = any, R = any>(): DialogContextValue<D, R> => {
+export const useDialogContext = <D = any, R = any>(): DialogContextValue<
+  D,
+  R
+> => {
   return useContext(IndividualDialogContext);
 };
