+++ packages/react-dialog-async/src/context/index.ts
@@ -1,0 +1,1 @@
+export { useDialogContext } from './IndividualDialogContext';
