+++ packages/react-dialog-async/src/index.ts
@@ -1,5 +1,6 @@
-export { default as DialogProvider } from './DialogProvider/DialogProvider';
-export { default as DialogOutlet } from './DialogOutlet';
-export { default as useDialog } from './useDialog';
-export { useDialogContext } from './IndividualDialogContext';
-export { type AsyncDialogProps } from './types';
+export { type AsyncDialogProps, type AsyncDialogComponent } from './types';
+
+export * from './context';
+export * from './useDialog';
+export * from './DialogProvider';
+export * from './DialogOutlet';
