+++ packages/react-dialog-async/src/useDialog/index.ts
@@ -1,0 +1,1 @@
+export { default as useDialog } from './useDialog';
